+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add position in frame to attendee updates
 
 ### Changed
 - Bump websocket-extensions from 0.1.3 to 0.1.4